added del_nonshm_lumps in sip.c
E-CSCF: fixed a bug: when the INVITE is forwarded to the PSAP, some lumps are added, but they are in pkg and when an error response was generated 
(e.g. 408 Request Timeout) or received, the function run_failure_handlers was considering the add_rm lump and trying to free it, because the flags 
of the memory chunck had invalid value (memory thread specific accessed from another thread), causing an error. Now the pkg lumps are deleted after 
relaying the request.